Rebasing : Added the functionality to propose a movie.
The logic is as follows:
The user proposes a movie
The server receives the proposition and checks whether the user has less than 3 movies already currently proposed.
If he has 3 or more movies, the client displays a modal to ask him this: either they cancel their proposition OR they choose to remove one of their current propositions
When this is done, the server receives the instruction and acts consequently. If one movie was chosen, it is removed from the "suggestions" table and the new one is added. If the proposition is canceled, nothing more happens.